+++ lib/src/domain/from_field_data.dart
@@ -5,21 +5,25 @@ class FormFieldData<T> {
   final String? error;
   final String? Function(T? value)? validator;
   final Key? key;
+  final Future<String?> Function(T? value)? asyncValidator;
 
-  FormFieldData( {
+  FormFieldData({
     T? initialValue,
     this.validator,
     this.error,
     this.key,
-  })  : value = ValueNotifier<T?>(initialValue);
+    this.asyncValidator,
+  }) : value = ValueNotifier<T?>(initialValue);
 
   @override
   bool operator ==(Object other) {
-    return other is FormFieldData<T> && other.value==value&&other.error==error;
+    return other is FormFieldData<T> &&
+        other.value == value &&
+        other.error == error;
   }
 
   @override
-  int get hashCode => value.hashCode^error.hashCode;
+  int get hashCode => value.hashCode ^ error.hashCode;
 
   /// Creates a copy of this [FormFieldData] with the ability to override some properties.
   FormFieldData<T> copyWith({
@@ -27,16 +31,17 @@ class FormFieldData<T> {
     String? error,
     String? Function(dynamic value)? validator,
     T? value,
+    Future<String?> Function(T? value)? asyncValidator,
   }) {
     return FormFieldData<T>(
       initialValue: initialValue ?? this.value.value,
       error: error ?? this.error,
       validator: validator ?? this.validator,
+      asyncValidator: asyncValidator ?? this.asyncValidator,
     );
   }
 
   void dispose() {
     value.dispose();
   }
-
 }